Logic for fetching and saving all issue related data
Creates public method `fetch_and_save!` that will ultimately get all closed issues, for all our repos.

Then calls for the organization of that data into a common, savable format, and saves it to the database as:

- GitHubStatistic records
- GitHubUser records

Smartly hits GitHub’s API for records that have been merged since the last record we have in our db, as well as only saving records to our db that do not already exist.